+++ src/Data/Geo/Jord/Transformation.hs
@@ -8,24 +8,24 @@
 --
 -- Coordinates transformation parameters.
 --
-module Data.Geo.Jord.Tx
+module Data.Geo.Jord.Transformation
     (
     -- * transformation parameters.
-      Tx(..)
-    , inverseTx
-    , TxParams(..)
-    , TxParams7
-    , TxRates
-    , TxParams15(..)
-    , txParams7
-    , txRates
-    , txParamsAt
+      Transformation(..)
+    , inverse
+    , Params(..)
+    , Params7
+    , Rates
+    , Params15(..)
+    , params7
+    , rates
+    , paramsAt
     -- * transformation graph.
-    , TxGraph
-    , txGraph
-    , txParamsBetween
+    , Graph
+    , graph
+    , paramsBetween
     -- * geocentric coordinate transformation
-    , transformGeoc
+    , apply
     ) where
 
 import Data.List (find, foldl', sortOn)
@@ -36,65 +36,61 @@ import qualified Data.Geo.Jord.Math3d as Math3d (add, multM, scale)
 import Data.Geo.Jord.Model (Epoch(..), ModelId)
 
 -- | Coordinate transformation between 2 models (A & B).
-data Tx a =
-    Tx
+data Transformation a =
+    Transformation
         { modelA :: ModelId -- ^  model A.
         , modelB :: ModelId -- ^ model B.
-        , txParams :: a -- ^ transformation parameters - i.e. 'modelA'-> 'modelB'
+        , params :: a -- ^ transformation parameters - i.e. 'modelA'-> 'modelB'
         }
 
 -- | inverse transformation.
-inverseTx :: (TxParams a) => Tx a -> Tx a
-inverseTx t = Tx (modelB t) (modelA t) (inverseTxParams (txParams t))
+inverse :: (Params a) => Transformation a -> Transformation a
+inverse t = Transformation (modelB t) (modelA t) (inverseParams (params t))
 
 -- | class for transformation parameters.
-class TxParams a where
-    idTxParams :: a -- ^ identity transformation parameters, i.e. @p T idTxParams = p@.
-    inverseTxParams :: a -> a -- ^ inverse transformation parameters.
+class Params a where
+    idParams :: a -- ^ identity transformation parameters, i.e. @p T idParams = p@.
+    inverseParams :: a -> a -- ^ inverse transformation parameters.
 
--- | 7-parameter transformation (Helmert); use 'txParams7' to construct.
-data TxParams7 =
-    TxParams7 !V3 !Double !V3
+-- | 7-parameter transformation (Helmert); use 'params7' to construct.
+data Params7 =
+    Params7 !V3 !Double !V3
     deriving (Show)
 
-instance TxParams TxParams7 where
-    idTxParams = TxParams7 (V3 0 0 0) 0 (V3 0 0 0)
-    inverseTxParams (TxParams7 c s r) =
-        TxParams7 (Math3d.scale c (-1.0)) (-s) (Math3d.scale r (-1.0))
+instance Params Params7 where
+    idParams = Params7 (V3 0 0 0) 0 (V3 0 0 0)
+    inverseParams (Params7 c s r) = Params7 (Math3d.scale c (-1.0)) (-s) (Math3d.scale r (-1.0))
 
-instance TxParams TxParams15 where
-    idTxParams = TxParams15 (Epoch 0) idTxParams (TxRates (V3 0 0 0) 0 (V3 0 0 0))
-    inverseTxParams (TxParams15 e p (TxRates c s r)) =
-        TxParams15
-            e
-            (inverseTxParams p)
-            (TxRates (Math3d.scale c (-1.0)) (-s) (Math3d.scale r (-1.0)))
+instance Params Params15 where
+    idParams = Params15 (Epoch 0) idParams (Rates (V3 0 0 0) 0 (V3 0 0 0))
+    inverseParams (Params15 e p (Rates c s r)) =
+        Params15 e (inverseParams p) (Rates (Math3d.scale c (-1.0)) (-s) (Math3d.scale r (-1.0)))
 
 -- | Transformation rates for the 15-parameter transformation (Helmert); use 'txRates' to construct.
-data TxRates =
-    TxRates !V3 !Double !V3
+data Rates =
+    Rates !V3 !Double !V3
     deriving (Show)
 
 -- | Epoch and 14-parameter transformation (Helmert).
-data TxParams15 =
-    TxParams15 Epoch TxParams7 TxRates
+data Params15 =
+    Params15 Epoch Params7 Rates
     deriving (Show)
 
 -- | 7-parameter transformation (Helmert) from given translation vector, scale factor and rotation matrix.
-txParams7 ::
+params7 ::
        (Double, Double, Double) -- ^ translation vector containing the three translations along the coordinate axes: tx, ty, tz in __millimetres__
     -> Double -- ^ scale factor (unitless) expressed in __parts per billion__
     -> (Double, Double, Double) -- ^ rotation matrix (orthogonal) consisting of the three axes rx, ry, rz in __milliarcseconds__
-    -> TxParams7
-txParams7 c s r = TxParams7 (mmToMetres c) (s / 1e9) (masToRadians r)
+    -> Params7
+params7 c s r = Params7 (mmToMetres c) (s / 1e9) (masToRadians r)
 
 -- | rates of the 15-parameter translation (Helmert) from given translation rates, scale factor rate and rotation rates.
-txRates ::
+rates ::
        (Double, Double, Double) -- ^ translation rate in __millimetres per year__.
     -> Double -- ^ scale factor rate in __part per billion per year__.
     -> (Double, Double, Double) -- ^ rotation rate in __milliarcseconds per year__.
-    -> TxRates
-txRates c s r = TxRates (mmToMetres c) (s / 1e9) (masToRadians r)
+    -> Rates
+rates c s r = Rates (mmToMetres c) (s / 1e9) (masToRadians r)
 
 mmToMetres :: (Double, Double, Double) -> V3
 mmToMetres (cx, cy, cz) = Math3d.scale (V3 cx cy cz) (1.0 / 1000.0)
@@ -104,9 +100,8 @@ masToRadians (rx, ry, rz) = Math3d.scale (V3 rx ry rz) (pi / (3600.0 * 1000.0 * 
 
 -- | @txParamsAt e tx15@ returns the 7-parameter transformation corresponding to the
 -- 15-parameter transformation @tx15@ at epoch @e@.
-txParamsAt :: Epoch -> TxParams15 -> TxParams7
-txParamsAt (Epoch e) (TxParams15 (Epoch pe) (TxParams7 c s r) (TxRates rc rs rr)) =
-    TxParams7 c' s' r'
+paramsAt :: Epoch -> Params15 -> Params7
+paramsAt (Epoch e) (Params15 (Epoch pe) (Params7 c s r) (Rates rc rs rr)) = Params7 c' s' r'
   where
     de = e - pe
     c' = Math3d.add c (Math3d.scale rc de)
@@ -132,40 +127,40 @@ data State =
     State [ModelId] [Path]
 
 -- | Transformation graph: vertices are 'ModelId' and edges are transformation parameters.
-data TxGraph a =
-    TxGraph ![Connection] ![Edge a]
+data Graph a =
+    Graph ![Connection] ![Edge a]
 
 -- | @txGraph ts@ returns a transformation graph containing all given direct and inverse
 -- (i.e. for each 'Tx': 'txParams' & 'inverseTxParams') transformations.
-txGraph :: (TxParams a) => [Tx a] -> TxGraph a
-txGraph = foldl' addTx emptyGraph
+graph :: (Params a) => [Transformation a] -> Graph a
+graph = foldl' addTx emptyGraph
 
 -- | @txParamsBetween m0 m1 g@ computes the ordered list of transformation parameters to be
 -- successively applied when transforming the coordinates of a position in model @m0@ to model @m1@.
 -- The returned list is empty, if either model is not in the graph (i.e. not a vertex)  or if no
 -- such transformation exists (i.e. model @m1@ cannot be reached from model @m0@).
-txParamsBetween :: (TxParams a) => ModelId -> ModelId -> TxGraph a -> [a]
-txParamsBetween m0 m1 g
-    | m0 == m1 = [idTxParams]
+paramsBetween :: (Params a) => ModelId -> ModelId -> Graph a -> [a]
+paramsBetween m0 m1 g
+    | m0 == m1 = [idParams]
     | null ms = []
     | otherwise = findParams ms g
   where
     ms = dijkstra (State [m0] []) m1 g
 
 -- | empty graph.
-emptyGraph :: TxGraph a
-emptyGraph = TxGraph [] []
+emptyGraph :: Graph a
+emptyGraph = Graph [] []
 
--- | add 'Tx' to graph.
-addTx :: (TxParams a) => TxGraph a -> Tx a -> TxGraph a
-addTx (TxGraph cs es) t = TxGraph cs' es'
+-- | add 'Transformation' to graph.
+addTx :: (Params a) => Graph a -> Transformation a -> Graph a
+addTx (Graph cs es) t = Graph cs' es'
   where
     ma = modelA t
     mb = modelB t
     cs1 = addConnection cs ma mb
     cs' = addConnection cs1 mb ma
-    txp = txParams t
-    es' = Edge ma txp mb : Edge mb (inverseTxParams txp) ma : es
+    txp = params t
+    es' = Edge ma txp mb : Edge mb (inverseParams txp) ma : es
 
 -- | add connection to graph.
 addConnection :: [Connection] -> ModelId -> ModelId -> [Connection]
@@ -184,8 +179,8 @@ addConnection cs m1 m2
     updated = cur {adjacents = m2 : adjacents cur}
 
 -- | successors of given model in given graph.
-successors :: ModelId -> TxGraph a -> [ModelId]
-successors m (TxGraph cs _) = concatMap adjacents (filter (\c -> node c == m) cs)
+successors :: ModelId -> Graph a -> [ModelId]
+successors m (Graph cs _) = concatMap adjacents (filter (\c -> node c == m) cs)
 
 -- | visit every given model from given model.
 visit :: ModelId -> [ModelId] -> State -> State
@@ -204,7 +199,7 @@ shortest c m ps = reverse (m : s)
     s = head (sortOn length fs)
 
 -- | dijkstra.
-dijkstra :: State -> ModelId -> TxGraph a -> [ModelId]
+dijkstra :: State -> ModelId -> Graph a -> [ModelId]
 dijkstra (State [] _) _ _ = []
 dijkstra (State [c] []) t g = dijkstra (State [c] [[c]]) t g
 dijkstra (State (c:r) v) t g
@@ -216,8 +211,8 @@ dijkstra (State (c:r) v) t g
     s'' = visit c succs s'
 
 -- | find tx params between given models: [A, B, C] => params (A, B), params (B, C).
-findParams :: [ModelId] -> TxGraph a -> [a]
-findParams ms (TxGraph _ es)
+findParams :: [ModelId] -> Graph a -> [a]
+findParams ms (Graph _ es)
     | length ps == length r = r
     | otherwise = []
   where
@@ -232,11 +227,10 @@ findParam p es = fmap (\(Edge _ pa _) -> pa) (find (edgeEq p) es)
 edgeEq :: (ModelId, ModelId) -> Edge a -> Bool
 edgeEq (m1, m2) (Edge m1' _ m2') = m1 == m1' && m2 == m2'
 
--- | @transformGeoc gc tx7@ returns the geocentric coordinates resulting from applying the 7-parameter
+-- | @apply gc tx7@ returns the geocentric coordinates resulting from applying the 7-parameter
 -- transformation @tx7@ to the geocentric coordinates represented by vector @gc@.
-transformGeoc :: V3 -> TxParams7 -> V3
-transformGeoc gc (TxParams7 c s r) =
-    Math3d.add c (Math3d.scale (Math3d.multM gc (rotation r)) (1.0 + s))
+apply :: V3 -> Params7 -> V3
+apply gc (Params7 c s r) = Math3d.add c (Math3d.scale (Math3d.multM gc (rotation r)) (1.0 + s))
 
 rotation :: V3 -> [V3]
 rotation (V3 x y z) = [V3 1.0 (-z) y, V3 z 1.0 (-x), V3 (-y) x 1.0]